refactor(analytic-logs): Replace TraceLogRepository with MessageSentCountsRepository and remove unused trace-event-counts schema
- Updated BuildMessagesDeliveredChart use case to utilize MessageSentCountsRepository for fetching message delivery data.
- Removed the trace-event-counts schema and related code from the repository as it is no longer needed.
- Cleaned up imports in the analytic-logs module to reflect the changes.
