+++ libs/application-generic/src/services/analytic-logs/message-sent-counts/index.ts
@@ -1,0 +1,2 @@
+export * from './message-sent-counts.repository';
+export * from './message-sent-counts.schema';
