+++ libs/application-generic/src/services/analytic-logs/trace-log/index.ts
@@ -1,3 +1,2 @@
 export * from './trace-log.repository';
 export * from './trace-log.schema';
-export * from './trace-event-counts.schema';
